Refactor. style 코드 분리 및 추가
